[DEV-5161] removing dead NAICS code
