+++ src/js/dataMapping/search/awardsOperationKeys.js
@@ -20,7 +20,6 @@ export const rootKeys = {
     awardID: 'award_ids',
     cfda: 'program_numbers',
     naics: 'naics_codes',
-    naics_v2: 'naics_codes',
     psc: 'psc_codes',
     contractPricing: 'contract_pricing_type_codes',
     setAsideType: 'set_aside_type_codes',
