+++ src/js/models/search/SearchAwardsOperation.js
@@ -43,7 +43,6 @@ class SearchAwardsOperation {
         this.selectedAwardIDs = [];
 
         this.selectedCFDA = [];
-        this.selectedNAICS = [];
         this.naicsCodes = checkboxTreeKeys;
         this.pscCheckbox = checkboxTreeKeys;
         this.pricingType = [];
@@ -87,7 +86,6 @@ class SearchAwardsOperation {
         this.selectedAwardIDs = state.selectedAwardIDs.toArray();
 
         this.selectedCFDA = state.selectedCFDA.toArray();
-        this.selectedNAICS = state.selectedNAICS.toArray();
         this.naicsCodes = {
             require: state.naicsCodes.toObject().require,
             exclude: state.naicsCodes.toObject().exclude
@@ -304,17 +302,12 @@ class SearchAwardsOperation {
         }
 
         // Add NAICS
-        if (this.selectedNAICS.length > 0) {
-            filters[rootKeys.naics] = this.selectedNAICS.map((naics) => naics.naics);
-        }
-
-        // NAICS v2
         if (this.naicsCodes.require.length > 0) {
             if (this.naicsCodes.exclude.length > 0) {
-                filters[rootKeys.naics_v2] = this.naicsCodes;
+                filters[rootKeys.naics] = this.naicsCodes;
             }
             else {
-                filters[rootKeys.naics_v2] = this.naicsCodes.require;
+                filters[rootKeys.naics] = this.naicsCodes.require;
             }
         }
