+++ tests/containers/search/filters/naics/mockNAICS.js
@@ -1,706 +1,806 @@
-import { List } from 'immutable';
+export const placeholderNodes = [
+    {
+        value: '11',
+        label: 'Agriculture, Forestry, Fishing and Hunting',
+        count: 64,
+        children: [
+            {
+                isPlaceHolder: true,
+                label: "Placeholder Value",
+                value: "children_of_11"
+            }
+        ]
+    },
+    {
+        value: '21',
+        label: 'Mining, Quarrying, and Oil and Gas Extraction',
+        count: 7,
+        children: [
+            {
+                isPlaceHolder: true,
+                label: "Placeholder Value",
+                value: "children_of_21"
+            }
+        ]
+    },
+    {
+        value: '54',
+        label: 'Professional, Scientific, and Technical Services',
+        count: 2,
+        children: [
+            {
+                isPlaceHolder: true,
+                label: "Placeholder Value",
+                value: "children_of_54"
+            }
+        ]
+    },
+    {
+        value: '71',
+        label: 'Arts, Entertainment, and Recreation',
+        count: 5,
+        children: [
+            {
+                isPlaceHolder: true,
+                label: "Placeholder Value",
+                value: "children_of_71"
+            }
+        ]
+    }
+];
 
-export const mockNAICS = {
-    results: [
+export const searchResults = [{
+    value: '11',
+    naics: '11',
+    label: 'Agriculture, Forestry, Fishing and Hunting',
+    count: 64,
+    children: [
         {
-            naics: "332993",
-            naics_description: "AMMUNITION (EXCEPT SMALL ARMS) MANUFACTURING"
-        },
-        {
-            naics: "333293",
-            naics_description: "PRINTING MACHINERY AND EQUIPMENT MANUFACTURING"
-        },
-        {
-            naics: "333313",
-            naics_description: "OFFICE MACHINERY MANUFACTURING"
-        },
-        {
-            naics: "333314",
-            naics_description: "OPTICAL INSTRUMENT AND LENS MANUFACTURING"
-        },
-        {
-            naics: "333315",
-            naics_description: "PHOTOGRAPHIC AND PHOTOCOPYING EQUIPMENT MANUFACTURING"
-        },
-        {
-            naics: "333316",
-            naics_description: "PHOTOGRAPHIC AND PHOTOCOPYING EQUIPMENT MANUFACTURING"
-        },
-        {
-            naics: "333318",
-            naics_description: "OTHER COMMERCIAL AND SERVICE INDUSTRY MACHINERY MANUFACTURING"
-        },
-        {
-            naics: "336111",
-            naics_description: "AUTOMOBILE MANUFACTURING"
-        },
-        {
-            naics: "339944",
-            naics_description: "CARBON PAPER AND INKED RIBBON MANUFACTURING"
-        },
-        {
-            naics: "339999",
-            naics_description: "ALL OTHER MISCELLANEOUS MANUFACTURING"
+            value: "1111",
+            naics: "1111",
+            naics_description: "Oilseed and Grain Farming",
+            count: 8,
+            children: [
+                {
+                    value: "111110",
+                    naics: "111110",
+                    naics_description: "Soybean Farming",
+                    count: 0
+                }
+            ]
         }
     ]
-};
+}];
 
-export const naicsMock2 = [
-    {
-        naics: '11',
-        naics_description: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        children: [
-            {
-                naics: '1111',
-                naics_description: 'Oilseed and Grain Farming',
-                count: 2,
-                children: [
-                    {
-                        naics: '111111',
-                        naics_description: 'More Oilseed Stuff and Grain Farming',
-                        count: 0
-                    },
-                    {
-                        naics: '111112',
-                        naics_description: 'Even More Oilseed Stuff and Grain Farming',
-                        count: 0
-                    }
-                ]
-            },
-            {
-                naics: '1112',
-                naics_description: 'Vegetable and Melon Farming',
-                count: 2,
-                children: [
-                    {
-                        naics: '111211',
-                        naics_description: 'Potato Farming',
-                        count: 0
-                    },
-                    {
-                        naics: '111219',
-                        naics_description: 'Other Vegetable (except Potato) and Melon Farming',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    },
-    {
-        naics: '21',
-        naics_description: 'Mining, Quarrying, and Oil and Gas Extraction',
-        count: 7,
-        children: [
-            {
-                naics: '2111',
-                naics_description: 'Oil and Gas Extraction',
-                count: 2,
-                children: [
-                    {
-                        naics: '211120',
-                        naics_description: 'Crude Petroleum Extraction',
-                        count: 0
-                    },
-                    {
-                        naics: '211130',
-                        naics_description: 'Natural Gas Extraction',
-                        count: 0
-                    }
-                ]
-            },
-            {
-                naics: '2121',
-                naics_description: 'Coal Mining',
-                count: 3,
-                children: [
-                    {
-                        naics: '212111',
-                        naics_description: 'Bituminous Coal and Lignite Surface Mining',
-                        count: 0
-                    },
-                    {
-                        naics: '212112',
-                        naics_description: 'Bituminous Coal Underground Mining',
-                        count: 0
-                    },
-                    {
-                        naics: '212113',
-                        naics_description: 'Anthracite Mining',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    },
-    {
-        naics: '54',
-        naics_description: 'Professional, Scientific, and Technical Services',
-        count: 2,
-        children: [
-            {
-                naics: '5411',
-                naics_description: 'Legal Services',
-                count: 0
-            },
-            {
-                naics: '5412',
-                naics_description: 'Accounting, Tax Preparation, Bookkeeping, and Payroll Services',
-                count: 0
-            }
-        ]
-    },
-    {
-        naics: '71',
-        naics_description: 'Arts, Entertainment, and Recreation',
-        count: 5,
-        children: [
-            {
-                naics: '7111',
-                naics_description: 'Performing Arts Companies',
-                count: 0
-            },
-            {
-                naics: '7112',
-                naics_description: 'Spectator Sports',
-                count: 3,
-                children: [
-                    {
-                        naics: '711211',
-                        naics_description: 'Sports Teams and Clubs',
-                        count: 0
-                    },
-                    {
-                        naics: '711212',
-                        naics_description: 'Racetracks',
-                        count: 0
-                    },
-                    {
-                        naics: '711219',
-                        naics_description: 'Other Spectator Sports',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    }
-];
-
-export const naicsMockCleanData = [
+export const treeWithPlaceholdersAndRealData = [
     {
         value: '11',
-        label: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        children: [
-            {
-                value: '1111',
-                label: 'Oilseed and Grain Farming',
-                count: 2,
-                children: [
-                    {
-                        value: '111111',
-                        label: 'More Oilseed Stuff and Grain Farming',
-                        count: 0
-                    },
-                    {
-                        value: '111112',
-                        label: 'Even More Oilseed Stuff and Grain Farming',
-                        count: 0
-                    }
-                ]
-            },
-            {
-                value: '1112',
-                label: 'Vegetable and Melon Farming',
-                count: 2,
-                children: [
-                    {
-                        value: '111211',
-                        label: 'Potato Farming',
-                        count: 0
-                    },
-                    {
-                        value: '111219',
-                        label: 'Other Vegetable (except Potato) and Melon Farming',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    },
-    {
-        value: '21',
-        label: 'Mining, Quarrying, and Oil and Gas Extraction',
-        count: 7,
-        children: [
-            {
-                value: '2111',
-                label: 'Oil and Gas Extraction',
-                count: 2,
-                children: [
-                    {
-                        value: '211120',
-                        label: 'Crude Petroleum Extraction',
-                        count: 0
-                    },
-                    {
-                        value: '211130',
-                        label: 'Natural Gas Extraction',
-                        count: 0
-                    }
-                ]
-            },
-            {
-                value: '2121',
-                label: 'Coal Mining',
-                count: 3,
-                children: [
-                    {
-                        value: '212111',
-                        label: 'Bituminous Coal and Lignite Surface Mining',
-                        count: 0
-                    },
-                    {
-                        value: '212112',
-                        label: 'Bituminous Coal Underground Mining',
-                        count: 0
-                    },
-                    {
-                        value: '212113',
-                        label: 'Anthracite Mining',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    },
-    {
-        value: '54',
-        label: 'Professional, Scientific, and Technical Services',
-        count: 2,
-        children: [
-            {
-                value: '5411',
-                label: 'Legal Services',
-                count: 0
-            },
-            {
-                value: '5412',
-                label: 'Accounting, Tax Preparation, Bookkeeping, and Payroll Services',
-                count: 0
-            }
-        ]
-    },
-    {
-        value: '71',
-        label: 'Arts, Entertainment, and Recreation',
-        count: 5,
-        children: [
-            {
-                value: '7111',
-                label: 'Performing Arts Companies',
-                count: 0
-            },
-            {
-                value: '7112',
-                label: 'Spectator Sports',
-                count: 3,
-                children: [
-                    {
-                        value: '711211',
-                        label: 'Sports Teams and Clubs',
-                        count: 0
-                    },
-                    {
-                        value: '711212',
-                        label: 'Racetracks',
-                        count: 0
-                    },
-                    {
-                        value: '711219',
-                        label: 'Other Spectator Sports',
-                        count: 0
-                    }
-                ]
-            }
-        ]
-    }
-];
-
-export const naicsMockInitialLoadApiResponse = [
-    {
         naics: '11',
         naics_description: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20
-    },
-    {
-        naics: '21',
-        naics_description: 'Mining, Quarrying, and Oil and Gas Extraction',
-        count: 7
-    },
-    {
-        naics: '54',
-        naics_description: 'Professional, Scientific, and Technical Services',
-        count: 2
-    },
-    {
-        naics: '71',
-        naics_description: 'Arts, Entertainment, and Recreation',
-        count: 5
+        count: 64,
+        children: [
+            {
+                value: "1111",
+                naics: "1111",
+                naics_description: "Oilseed and Grain Farming",
+                count: 8,
+                children: [
+                    {
+                        value: "111110",
+                        naics: "111110",
+                        naics_description: "Soybean Farming",
+                        count: 0
+                    },
+                    {
+                        value: "111120",
+                        naics: "111120",
+                        naics_description: "Oilseed (except Soybean) Farming",
+                        count: 0
+                    },
+                    {
+                        value: "children_of_1111",
+                        naics: "children_of_1111",
+                        isPlaceHolder: true,
+                        count: 0
+                    }
+                ]
+            },
+            {
+                value: "children_of_11",
+                naics: "children_of_11",
+                isPlaceHolder: true,
+                count: 0
+            }
+        ]
     }
 ];
 
-export const naicsMockCleanDataInitialLoad = [
+export const treeWithPlaceholdersAndRealDataPSCDepth = [
     {
         value: '11',
-        label: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        children: [{}],
-        path: [0]
-    },
-    {
-        value: '21',
-        label: 'Mining, Quarrying, and Oil and Gas Extraction',
-        count: 7,
-        children: [{}],
-        path: [1]
-    },
-    {
-        value: '54',
-        label: 'Professional, Scientific, and Technical Services',
-        count: 2,
-        children: [{}],
-        path: [2]
-    },
-    {
-        value: '71',
-        label: 'Arts, Entertainment, and Recreation',
-        count: 5,
-        children: [{}],
-        path: [3]
+        naics: '11',
+        naics_description: 'Agriculture, Forestry, Fishing and Hunting',
+        count: 64,
+        children: [
+            {
+                value: "1111",
+                naics: "1111",
+                naics_description: "Oilseed and Grain Farming",
+                count: 8,
+                children: [
+                    {
+                        naics: "111110",
+                        value: "111110",
+                        naics_description: "Soybean Farming",
+                        count: 0
+                    },
+                    {
+                        value: "111120",
+                        naics: "111120",
+                        naics_description: "Oilseed (except Soybean) Farming",
+                        count: 20,
+                        children: [
+                            {
+                                naics: 'dont-overwrite-me',
+                                value: 'dont-overwrite-me',
+                                naics_description: 'real'
+                            },
+                            {
+                                naics: 'children_of_111120',
+                                value: 'children_of_111120',
+                                naics_description: 'psc-depth not realistic naics data',
+                                isPlaceHolder: true
+                            }
+                        ]
+                    },
+                    {
+                        value: "children_of_1111",
+                        naics: "children_of_1111",
+                        isPlaceHolder: true,
+                        count: 0
+                    }
+                ]
+            },
+            {
+                value: "children_of_11",
+                naics: "children_of_11",
+                isPlaceHolder: true,
+                count: 0
+            }
+        ]
     }
 ];
 
-export const naicsMockAPIResponse = [
+export const reallyBigTree = [
     {
-        naics: '11',
-        naics_description: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
+        naics: "11",
+        value: "11",
+        naics_description: "Agriculture, Forestry, Fishing and Hunting",
+        count: 64,
         children: [
             {
-                naics: '1111',
-                naics_description: 'Agriculture, Fishing and Hunting',
-                count: 20
+                naics: "1111",
+                value: "1111",
+                naics_description: "Oilseed and Grain Farming",
+                count: 8,
+                children: [
+                    {
+                        naics: "111110",
+                        value: "111110",
+                        naics_description: "Soybean Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111120",
+                        value: "111120",
+                        naics_description: "Oilseed (except Soybean) Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111130",
+                        value: "111130",
+                        naics_description: "Dry Pea and Bean Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111140",
+                        value: "111140",
+                        naics_description: "Wheat Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111150",
+                        value: "111150",
+                        naics_description: "Corn Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111160",
+                        value: "111160",
+                        naics_description: "Rice Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111191",
+                        value: "111191",
+                        naics_description: "Oilseed and Grain Combination Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111199",
+                        value: "111199",
+                        naics_description: "All Other Grain Farming",
+                        count: 0
+                    }
+                ]
             },
             {
-                naics: '1121',
-                naics_description: 'Mining, and Oil and Gas Extraction',
-                count: 7
-            },
-            {
-                naics: '1154',
-                naics_description: 'Professional, and Technical Services',
+                naics: "1112",
+                value: "1112",
+                naics_description: "Vegetable and Melon Farming",
                 count: 2
             },
             {
-                naics: '1171',
-                naics_description: 'Arts, and Recreation',
+                naics: "1113",
+                value: "1113",
+                naics_description: "Fruit and Tree Nut Farming",
+                count: 9
+            },
+            {
+                naics: "1114",
+                value: "1114",
+                naics_description: "Greenhouse, Nursery, and Floriculture Production",
+                count: 4
+            },
+            {
+                naics: "1119",
+                value: "1119",
+                naics_description: "Other Crop Farming",
+                count: 7
+            },
+            {
+                naics: "1121",
+                value: "1121",
+                naics_description: "Cattle Ranching and Farming",
+                count: 4
+            },
+            {
+                naics: "1122",
+                value: "1122",
+                naics_description: "Hog and Pig Farming",
+                count: 1
+            },
+            {
+                naics: "1123",
+                value: "1123",
+                naics_description: "Poultry and Egg Production",
                 count: 5
+            },
+            {
+                naics: "1124",
+                value: "1124",
+                naics_description: "Sheep and Goat Farming",
+                count: 2
+            },
+            {
+                naics: "1125",
+                value: "1125",
+                naics_description: "Aquaculture",
+                count: 3
+            },
+            {
+                naics: "1129",
+                value: "1129",
+                naics_description: "Other Animal Production",
+                count: 4
+            },
+            {
+                naics: "1131",
+                value: "1131",
+                naics_description: "Timber Tract Operations",
+                count: 1
+            },
+            {
+                naics: "1132",
+                value: "1132",
+                naics_description: "Forest Nurseries and Gathering of Forest Products",
+                count: 1
+            },
+            {
+                naics: "1133",
+                value: "1133",
+                naics_description: "Logging",
+                count: 1
+            },
+            {
+                naics: "1141",
+                value: "1141",
+                naics_description: "Fishing",
+                count: 3
+            },
+            {
+                naics: "1142",
+                value: "1142",
+                naics_description: "Hunting and Trapping",
+                count: 1
+            },
+            {
+                naics: "1151",
+                value: "1151",
+                naics_description: "Support Activities for Crop Production",
+                count: 6
+            },
+            {
+                naics: "1152",
+                value: "1152",
+                naics_description: "Support Activities for Animal Production",
+                count: 1
+            },
+            {
+                naics: "1153",
+                value: "1153",
+                naics_description: "Support Activities for Forestry",
+                count: 1,
+                children: [
+                    {
+                        naics: "115310",
+                        value: "115310",
+                        naics_description: "Support Activities for Forestry",
+                        count: 0,
+                        className: 'hide'
+                    }
+                ]
             }
         ]
+    },
+    {
+        naics: "21",
+        value: "21",
+        naics_description: "Mining, Quarrying, and Oil and Gas Extraction",
+        count: 32
+    },
+    {
+        naics: "22",
+        naics_description: "Utilities",
+        count: 15
+    },
+    {
+        naics: "23",
+        naics_description: "Construction",
+        count: 31
+    },
+    {
+        naics: "31",
+        naics_description: "Manufacturing",
+        count: 133
+    },
+    {
+        naics: "32",
+        naics_description: "Manufacturing",
+        count: 136
+    },
+    {
+        naics: "33",
+        naics_description: "Manufacturing",
+        count: 269
+    },
+    {
+        naics: "42",
+        naics_description: "Wholesale Trade",
+        count: 71
+    },
+    {
+        naics: "44",
+        naics_description: "Retail Trade",
+        count: 50
+    },
+    {
+        naics: "45",
+        naics_description: "Retail Trade",
+        count: 33
+    },
+    {
+        naics: "48",
+        naics_description: "Transportation and Warehousing",
+        count: 50
+    },
+    {
+        naics: "49",
+        naics_description: "Transportation and Warehousing",
+        count: 7
+    },
+    {
+        naics: "51",
+        naics_description: "Information",
+        count: 43
+    },
+    {
+        naics: "52",
+        naics_description: "Finance and Insurance",
+        count: 42
+    },
+    {
+        naics: "53",
+        naics_description: "Real Estate and Rental and Leasing",
+        count: 29
+    },
+    {
+        naics: "54",
+        naics_description: "Professional, Scientific, and Technical Services",
+        count: 52
+    },
+    {
+        naics: "55",
+        naics_description: "Management of Companies and Enterprises",
+        count: 3
+    },
+    {
+        naics: "56",
+        naics_description: "Administrative and Support and Waste Management and Remediation Services",
+        count: 45
+    },
+    {
+        naics: "61",
+        naics_description: "Educational Services",
+        count: 17
+    },
+    {
+        naics: "62",
+        naics_description: "Health Care and Social Assistance",
+        count: 39
+    },
+    {
+        naics: "71",
+        naics_description: "Arts, Entertainment, and Recreation",
+        count: 25
+    },
+    {
+        naics: "72",
+        naics_description: "Accommodation and Food Services",
+        count: 19
+    },
+    {
+        naics: "81",
+        naics_description: "Other Services (except Public Administration)",
+        count: 49
+    },
+    {
+        naics: "92",
+        naics_description: "Public Administration",
+        count: 29
     }
 ];
 
-export const naicsMockAPIResponseClean = [
+export const reallyBigTreePSCDepth = [
     {
-        value: '11',
-        label: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        path: [0],
+        naics: "11",
+        value: "11",
+        naics_description: "Agriculture, Forestry, Fishing and Hunting",
+        count: 64,
         children: [
             {
-                value: '1111',
-                label: 'Agriculture, Fishing and Hunting',
-                count: 20,
-                path: [0, 0],
-                children: [{}]
+                naics: "1111",
+                value: "1111",
+                naics_description: "Oilseed and Grain Farming",
+                count: 8,
+                children: [
+                    {
+                        naics: "111110",
+                        value: "111110",
+                        naics_description: "Soybean Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111120",
+                        value: "111120",
+                        naics_description: "Oilseed (except Soybean) Farming",
+                        count: 1,
+                        children: [{
+                            naics: 'great grandchild',
+                            value: 'great grandchild',
+                            naics_description: 'This will never happen with NAICS; its a test for PSC',
+                            className: 'hide'
+                        }]
+                    },
+                    {
+                        naics: "111130",
+                        value: "111130",
+                        naics_description: "Dry Pea and Bean Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111140",
+                        value: "111140",
+                        naics_description: "Wheat Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111150",
+                        value: "111150",
+                        naics_description: "Corn Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111160",
+                        value: "111160",
+                        naics_description: "Rice Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111191",
+                        value: "111191",
+                        naics_description: "Oilseed and Grain Combination Farming",
+                        count: 0
+                    },
+                    {
+                        naics: "111199",
+                        value: "111199",
+                        naics_description: "All Other Grain Farming",
+                        count: 0
+                    }
+                ]
             },
             {
-                value: '1121',
-                label: 'Mining, and Oil and Gas Extraction',
-                count: 7,
-                path: [0, 1],
-                children: [{}]
+                naics: "1112",
+                value: "1112",
+                naics_description: "Vegetable and Melon Farming",
+                count: 2
             },
             {
-                value: '1154',
-                label: 'Professional, and Technical Services',
-                count: 2,
-                path: [0, 2],
-                children: [{}]
+                naics: "1113",
+                value: "1113",
+                naics_description: "Fruit and Tree Nut Farming",
+                count: 9
             },
             {
-                value: '1171',
-                label: 'Arts, and Recreation',
-                count: 5,
-                path: [0, 3],
-                children: [{}]
+                naics: "1114",
+                value: "1114",
+                naics_description: "Greenhouse, Nursery, and Floriculture Production",
+                count: 4
+            },
+            {
+                naics: "1119",
+                value: "1119",
+                naics_description: "Other Crop Farming",
+                count: 7
+            },
+            {
+                naics: "1121",
+                value: "1121",
+                naics_description: "Cattle Ranching and Farming",
+                count: 4
+            },
+            {
+                naics: "1122",
+                value: "1122",
+                naics_description: "Hog and Pig Farming",
+                count: 1
+            },
+            {
+                naics: "1123",
+                value: "1123",
+                naics_description: "Poultry and Egg Production",
+                count: 5
+            },
+            {
+                naics: "1124",
+                value: "1124",
+                naics_description: "Sheep and Goat Farming",
+                count: 2
+            },
+            {
+                naics: "1125",
+                value: "1125",
+                naics_description: "Aquaculture",
+                count: 3
+            },
+            {
+                naics: "1129",
+                value: "1129",
+                naics_description: "Other Animal Production",
+                count: 4
+            },
+            {
+                naics: "1131",
+                value: "1131",
+                naics_description: "Timber Tract Operations",
+                count: 1
+            },
+            {
+                naics: "1132",
+                value: "1132",
+                naics_description: "Forest Nurseries and Gathering of Forest Products",
+                count: 1
+            },
+            {
+                naics: "1133",
+                value: "1133",
+                naics_description: "Logging",
+                count: 1
+            },
+            {
+                naics: "1141",
+                value: "1141",
+                naics_description: "Fishing",
+                count: 3
+            },
+            {
+                naics: "1142",
+                value: "1142",
+                naics_description: "Hunting and Trapping",
+                count: 1
+            },
+            {
+                naics: "1151",
+                value: "1151",
+                naics_description: "Support Activities for Crop Production",
+                count: 6
+            },
+            {
+                naics: "1152",
+                value: "1152",
+                naics_description: "Support Activities for Animal Production",
+                count: 1
+            },
+            {
+                naics: "1153",
+                value: "1153",
+                naics_description: "Support Activities for Forestry",
+                count: 1,
+                children: [
+                    {
+                        naics: "115310",
+                        value: "115310",
+                        naics_description: "Support Activities for Forestry",
+                        count: 0,
+                        className: 'hide'
+                    }
+                ]
             }
         ]
-    }
-];
-
-export const naicsMockDataDeepDirty = [
+    },
     {
-        naics: '11',
-        naics_description: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        children: [
-            {
-                naics: '1111',
-                naics_description: 'Agriculture, Fishing and Hunting',
-                count: 20,
-                children: [
-                    {
-                        naics: '111111',
-                        naics_description: 'Agriculture, Fishing',
-                        count: 20,
-                        children: [
-                            {
-                                naics: '11111111',
-                                naics_description: 'Agriculture',
-                                count: 20
-                            },
-                            {
-                                naics: '11111121',
-                                naics_description: 'Agricu',
-                                count: 20
-                            }
-                        ]
-                    },
-                    {
-                        naics: '111121',
-                        naics_description: 'Agriculture, Fishing 2',
-                        count: 20,
-                        children: [
-                            {
-                                naics: '111111',
-                                naics_description: 'Agriculture 2',
-                                count: 20
-                            },
-                            {
-                                naics: '111111',
-                                naics_description: 'Agricu 2',
-                                count: 20
-                            }
-                        ]
-                    }
-                ]
-            },
-            {
-                naics: '1121',
-                naics_description: 'Mining, and Oil and Gas Extraction',
-                count: 7,
-                children: [
-                    {
-                        naics: '112111',
-                        naics_description: 'Mining, and Oil and Gas',
-                        count: 7,
-                        children: [
-                            {
-                                naics: '11211111',
-                                naics_description: 'Mining, and Oil',
-                                count: 7,
-                                children: [
-                                    {
-                                        naics: '11211111',
-                                        naics_description: 'Mining, and Oil',
-                                        count: 7
-                                    },
-                                    {
-                                        naics: '11211111',
-                                        naics_description: 'Mining, and Oil',
-                                        count: 7
-                                    }
-                                ]
-                            },
-                            {
-                                naics: '11211121',
-                                naics_description: 'Mining 3',
-                                count: 7,
-                                children: [
-                                    {
-                                        naics: '1121112111',
-                                        naics_description: 'Mining Oil',
-                                        count: 7
-                                    },
-                                    {
-                                        naics: '1121112121',
-                                        naics_description: 'Mining',
-                                        count: 7
-                                    }
-                                ]
-                            }
-                        ]
-                    },
-                    {
-                      
-                    }
-                ]
-            }
-        ]
-    }
-];
-
-export const naicsMockDataDeepClean = [
+        naics: "21",
+        value: "21",
+        naics_description: "Mining, Quarrying, and Oil and Gas Extraction",
+        count: 32
+    },
     {
-        value: '11',
-        label: 'Agriculture, Forestry, Fishing and Hunting',
-        count: 20,
-        path: [0],
-        children: [
-            {
-                value: '1111',
-                label: 'Agriculture, Fishing and Hunting',
-                count: 20,
-                path: [0, 1],
-                children: [
-                    {
-                        value: '111111',
-                        label: 'Agriculture, Fishing',
-                        count: 20,
-                        path: [0, 1, 0],
-                        children: [
-                            {
-                                value: '11111111',
-                                label: 'Agriculture',
-                                count: 20,
-                                path: [0, 1, 0, 0]
-                            },
-                            {
-                                value: '11111121',
-                                label: 'Agricu',
-                                count: 20,
-                                path: [0, 1, 0, 1]
-                            }
-                        ]
-                    },
-                    {
-                        value: '111121',
-                        label: 'Agriculture, Fishing 2',
-                        count: 20,
-                        path: [0, 1, 1],
-                        children: [
-                            {
-                                value: '111111',
-                                label: 'Agriculture 2',
-                                count: 20,
-                                path: [0, 1, 1, 0]
-                            },
-                            {
-                                value: '111111',
-                                label: 'Agricu 2',
-                                count: 20,
-                                path: [0, 1, 1, 1]
-                            }
-                        ]
-                    }
-                ]
-            },
-            {
-                value: '1121',
-                label: 'Mining, and Oil and Gas Extraction',
-                count: 7,
-                path: [0, 1],
-                children: [
-                    {
-                        value: '112111',
-                        label: 'Mining, and Oil and Gas',
-                        count: 7,
-                        path: [0, 1, 0],
-                        children: [
-                            {
-                                value: '11211111',
-                                label: 'Mining, and Oil',
-                                count: 7,
-                                path: [0, 1, 0, 1],
-                                children: [
-                                    {
-                                        value: '11211111',
-                                        label: 'Mining, and Oil',
-                                        count: 7,
-                                        path: [0, 1, 0, 1, 0]
-                                    },
-                                    {
-                                        value: '11211111',
-                                        label: 'Mining, and Oil',
-                                        count: 7,
-                                        path: [0, 1, 0, 1, 1]
-                                    }
-                                ]
-                            },
-                            {
-                                value: '11211121',
-                                label: 'Mining 3',
-                                count: 7,
-                                path: [0, 1, 0, 2],
-                                children: [
-                                    {
-                                        value: '1121112111',
-                                        label: 'Mining Oil',
-                                        count: 7,
-                                        path: [0, 1, 0, 2, 0]
-                                    },
-                                    {
-                                        value: '1121112121',
-                                        label: 'Mining',
-                                        count: 7,
-                                        path: [0, 1, 0, 2, 1]
-                                    }
-                                ]
-                            }
-                        ]
-                    },
-                    {
-                      
-                    }
-                ]
-            }
-        ]
+        naics: "22",
+        naics_description: "Utilities",
+        count: 15
+    },
+    {
+        naics: "23",
+        naics_description: "Construction",
+        count: 31
+    },
+    {
+        naics: "31",
+        naics_description: "Manufacturing",
+        count: 133
+    },
+    {
+        naics: "32",
+        naics_description: "Manufacturing",
+        count: 136
+    },
+    {
+        naics: "33",
+        naics_description: "Manufacturing",
+        count: 269
+    },
+    {
+        naics: "42",
+        naics_description: "Wholesale Trade",
+        count: 71
+    },
+    {
+        naics: "44",
+        naics_description: "Retail Trade",
+        count: 50
+    },
+    {
+        naics: "45",
+        naics_description: "Retail Trade",
+        count: 33
+    },
+    {
+        naics: "48",
+        naics_description: "Transportation and Warehousing",
+        count: 50
+    },
+    {
+        naics: "49",
+        naics_description: "Transportation and Warehousing",
+        count: 7
+    },
+    {
+        naics: "51",
+        naics_description: "Information",
+        count: 43
+    },
+    {
+        naics: "52",
+        naics_description: "Finance and Insurance",
+        count: 42
+    },
+    {
+        naics: "53",
+        naics_description: "Real Estate and Rental and Leasing",
+        count: 29
+    },
+    {
+        naics: "54",
+        naics_description: "Professional, Scientific, and Technical Services",
+        count: 52
+    },
+    {
+        naics: "55",
+        naics_description: "Management of Companies and Enterprises",
+        count: 3
+    },
+    {
+        naics: "56",
+        naics_description: "Administrative and Support and Waste Management and Remediation Services",
+        count: 45
+    },
+    {
+        naics: "61",
+        naics_description: "Educational Services",
+        count: 17
+    },
+    {
+        naics: "62",
+        naics_description: "Health Care and Social Assistance",
+        count: 39
+    },
+    {
+        naics: "71",
+        naics_description: "Arts, Entertainment, and Recreation",
+        count: 25
+    },
+    {
+        naics: "72",
+        naics_description: "Accommodation and Food Services",
+        count: 19
+    },
+    {
+        naics: "81",
+        naics_description: "Other Services (except Public Administration)",
+        count: 49
+    },
+    {
+        naics: "92",
+        naics_description: "Public Administration",
+        count: 29
     }
 ];
 
-export const emptyNAICSRedux = {
-    naics: new List(),
-    expanded: new List(),
-    checked: new List()
-};
-
-export const populatedNAICSRedux = {
-    naics: new List(naicsMockCleanData),
-    expanded: new List(['11', '21']),
-    checked: new List()
-};
-
-export const emptyNAICSProps = {
-    updateSelectedNAICS: () => {},
-    selectedNAICS: {},
-    appliedNAICS: {},
-    setNaics: () => {},
+export const defaultProps = {
+    expanded: [],
+    checked: [],
+    unchecked: [],
+    counts: [],
+    nodes: placeholderNodes,
+    searchExpanded: [],
+    stageNaics: () => {},
+    setNaicsNodes: () => {},
     setExpanded: () => {},
-    nodes: new List(),
-    expanded: new List(),
-    checked: new List()
-};
-
-export const populatedState = {
-    naics: naicsMockCleanData,
-    expanded: ['11', '12'],
-    checked: ['1111'],
-    isError: false,
-    errorMessage: '',
-    isLoading: false,
-    isSearch: false,
-    requestType: '',
-    fromRedux: false
+    setCheckedNaics: () => {},
+    removeNAICS: () => {},
+    setSearchedNaics: () => {},
+    setNaicsCounts: () => {},
+    addCheckedNaics: () => {},
+    showNaicsTree: () => {},
+    setUncheckedNaics: () => {},
+    uncheckedFromHash: [],
+    checkedFromHash: [],
+    countsFromHash: [],
+    filters: {
+        naicsCodes: {
+            require: [],
+            exclude: []
+        }
+    }
 };
