+++ tests/containers/search/mockSearchHashes.js
@@ -26,14 +26,13 @@ export const mockFilters = {
             selectedRecipientLocations: {},
             awardAmounts: {},
             timePeriodEnd: null,
-            naicsCodes: { require: [], exclude: [], counts: [] },
             selectedCFDA: {},
-            selectedNAICS: {},
             pricingType: [],
             setAside: [],
             extentCompeted: [],
             federalAccounts: [],
             treasuryAccounts: [],
+            naicsCodes: { require: [], exclude: [], counts: [] },
             tasCodes: { require: [], exclude: [], counts: [] },
             pscCodes: { require: [], exclude: [], counts: [] },
             defCodes: { require: [], exclude: [], counts: [] }
