+++ tests/helpers/naicsHelper-test.js
@@ -4,7 +4,7 @@ import {
     getImmediateAncestorNaicsCode
 } from 'helpers/naicsHelper';
 
-import * as mockData from '../containers/search/filters/naics/mockNaics_v2';
+import * as mockData from '../containers/search/filters/naics/mockNaics';
 
 describe('naicsHelper', () => {
     describe('getNaicsNodeFromTree', () => {
